backend/requirements.txt
fastapi
uvicorn
sqlalchemy[asyncio]
alembic
pydantic
passlib
bcrypt
python-jose
boto3
sendgrid
asyncpg
pytest
httpx
celery[redis]
pytest-asyncio


